Move the remaining widgets of the ui module into the core namespace
BaseImage, CropImage, CounterSet, CounterSetItem and DirectoryFilters join
the other core widgets, together with their three views.

DirectoryFilters stays a top-level widget rather than moving next to the
filter widgets still in the module. The two are unrelated: DirectoryFilters
extends humhub\components\Widget directly and drives the directory pages,
while the filter widgets build the stream filter panel. Grouping them now
would prejudge what happens to that folder.

Every old class name remains as a deprecated subclass until 1.21.
