+++ protected/humhub/modules/marketplace/widgets/ModuleFilters.php
@@ -12,7 +12,7 @@ use humhub\helpers\Html;
 use humhub\modules\admin\libs\HumHubAPI;
 use humhub\modules\marketplace\Module;
 use humhub\widgets\Icon;
-use humhub\modules\ui\widgets\DirectoryFilters;
+use humhub\widgets\DirectoryFilters;
 use humhub\widgets\bootstrap\Link;
 use Yii;
 use yii\helpers\Url;
