+++ protected/humhub/modules/space/widgets/Image.php
@@ -10,7 +10,7 @@ namespace humhub\modules\space\widgets;
 
 use humhub\helpers\Html;
 use humhub\modules\space\models\Space;
-use humhub\modules\ui\widgets\BaseImage;
+use humhub\widgets\BaseImage;
 
 /**
  * Return space image or acronym
