+++ protected/humhub/modules/ui/widgets/CounterSet.php
@@ -2,39 +2,15 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
 namespace humhub\modules\ui\widgets;
 
-use humhub\components\Widget;
-
 /**
- * Class CounterSet
- *
- * @since 1.3
- * @package humhub\modules\ui\widgets
+ * @deprecated since 1.20, use \humhub\widgets\CounterSet instead. Will be removed in 1.21.
  */
-class CounterSet extends Widget
+class CounterSet extends \humhub\widgets\CounterSet
 {
-    /**
-     * @var CounterSetItem[]
-     */
-    public $counters = [];
-
-
-    /**
-     * @var string the template to use
-     */
-    public $template = '@ui/widgets/views/counterSetHeader';
-
-
-    /**
-     * @inheritdoc
-     */
-    public function run()
-    {
-        return $this->render($this->template, ['counters' => $this->counters]);
-    }
 }
