+++ protected/humhub/modules/user/widgets/Image.php
@@ -9,7 +9,7 @@
 namespace humhub\modules\user\widgets;
 
 use humhub\helpers\Html;
-use humhub\modules\ui\widgets\BaseImage;
+use humhub\widgets\BaseImage;
 use humhub\modules\user\models\User;
 use humhub\modules\user\services\IsOnlineService;
 use Yii;
