+++ protected/humhub/modules/user/widgets/ProfileHeaderCounterSet.php
@@ -11,8 +11,8 @@ namespace humhub\modules\user\widgets;
 use humhub\modules\friendship\models\Friendship;
 use humhub\modules\space\models\Membership;
 use humhub\modules\space\models\Space;
-use humhub\modules\ui\widgets\CounterSetItem;
-use humhub\modules\ui\widgets\CounterSet;
+use humhub\widgets\CounterSetItem;
+use humhub\widgets\CounterSet;
 use humhub\modules\user\models\User;
 use Yii;
 use yii\helpers\Url;
