+++ protected/humhub/widgets/views/directoryFilters.php
@@ -6,7 +6,7 @@
  */
 
 use humhub\helpers\Html;
-use humhub\modules\ui\widgets\DirectoryFilters;
+use humhub\widgets\DirectoryFilters;
 use yii\helpers\Url;
 
 /* @var $directoryFilters DirectoryFilters */
